+++ sphinx/changelog.rst
@@ -14,6 +14,7 @@ Version 3.0
 - Switch patterns management to long format everywhere
 - Change the format of human readable patterns to a more readable markdown export
 - Implement parallelisation for finding patterns and finding applicable patterns. See `cpus` config option.
+- Change management of Metadata class. Outputs from Qumin are now shipped as a `Frictionless DataPackage <https://datapackage.org/>`_.
 - Change management of sampling:
     - `most_freq=False` replaced by `force_random=True`.
     - Addition of a seed option to determinise sampling.  By default, sample by frequency.